Increased implementation, reconsidering design
The code in test/main.c describes an API that is in some ways more
complicated than necessary.

TODO: Instead of having both update and delete buffers try to track
exactly what has changed, represent the idea of a buffer (or even a
column) that is paired with indices into another column of possibly
larger but at least equal size.