Custom memory allocation
This is probably premature, but I just did a bit of research into memory
management in C and it turns out this is a pretty common practice.
Since this is a library, it should attempt to handle out of memory
errors gracefully, leaving structures in a coherent state where
possible, and returning error values.