Remove ent_model
This abstraction was inappropriate because it suggested that the scope
of a processor or session should be limited to one set of tables and
that each table can exist in only one such set.  This was at odds with
the idea that processors can be defined over arbitrary sets of tables so
that all kinds of work can be done within the scope of a single session.